[#72] feat: patch handler, delete handler 하위 컴포넌트로 전달하여 사용하도록 변경
- get handler 분리하여 패치와 딜리트 이벤트 일어날 때마다 적용 되도록 함
